requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioredis==2.0.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.1
black==23.3.0
cachetools==5.3.0
celery==5.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comtypes==1.1.14
cowsay==5.0
cryptography==43.0.3
cs50==9.2.5
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==65.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==6.1.3
exceptiongroup==1.1.2
fastapi==0.115.2
Flask==2.2.3
frozenlist==1.5.0
ghp-import==2.1.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.21.0
httpx==0.27.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.4.2
Markdown==3.4.4
MarkupSafe==2.1.2
mergedeep==1.3.4
mkdocs==1.5.1
multidict==6.1.0
mypy==1.3.0
mypy-extensions==1.0.0
networkx==3.4.1
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
postgres==4.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.22.0
psycopg2-binary==2.9.6
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.0.9
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.12
pyttsx3==2.90
pytube==15.0.0
pywin32==305
PyYAML==6.0.1
pyyaml_env_tag==0.1
redis==5.2.0
requests==2.28.2
rsa==4.9
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
starlette==0.39.2
termcolor==2.2.0
trio==0.22.1
trio-websocket==0.10.3
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.31.1
vine==5.1.0
watchdog==3.0.0
wcwidth==0.2.13
websocket-client==1.6.1
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.16.0
